Add embedSourcemap option and description
This is one of the API options of both sass and node-sass. I needed it and had to look for it in the sass package API documentation. Adding it to the Readme will be useful.

It may also be useful mentioning that options are passed on to the node-sass / sass package directly, and so all API options are available. sass and node-sass have identical APIs, the intent being that they are interchange-able.